fix: improve error handling
